Refactored code and file structure
separated UI and logic
