+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   path_provider:
   path:
   intl: ^0.18.1
+  provider: ^6.0.5
   dotted_border: ^3.0.0
   material_symbols_icons: ^4.2661.0
   # The following adds the Cupertino Icons font to your application.
